mud-connections: Handle mud and char gconf key path retrieval properly
Instead of taking the visible name and parsing that, explicitly put the gconf path
components into the list store as columns, so we don't need to try to calculate the
path, but always get the correct thing as it was read from to create this liststore
item in the first place.

This fixes reading connection data from connections created in v0.11.2 or before
for connections that contain whitespace in MUD or character name, because the new
code escaped it from the display name differently than old. Old stripped whitespace,
new used gconf_escape_key, which instead of stripping converts a space to %32%.
By just carrying the key paths in the list store, we don't have any problems with
that, and it feels cleaner as well, even into GSettings conversion later on.

New connections get their gconf keys escaped instead of whitespace stripped, but
this will be fine as well on upgrade - possibly not after downgrade though.