Port to meson build system
All config.h includes are removed, as it is included for all via
-include preprocessor argument now.
Autotools is dropped completely, except for the unused
gnome-mud-manual bits, as the rest is remaining there too for now.
Outdated ChangeLog is removed.
Various autotools standard toplevel files are removed, some others
are updated.

As a result, we don't depend on intltool anymore, either.