Refactor and improve mud input box handling
Introduce a MudInputView class, which is a GtkStack based composite
widget, that contains the GtkTextView and GtkEntry, allowing to start
hiding the internals inside this new class. Now using a GtkStack
allows to use some animations and otherwise conceptually better
handling of this double widget thing from lack of GtkTextView
supporting password entry. It also simplifies all the grab_focus
calls, as this separate class now overrides grab_focus and passes
it on to the active child, so the caller doesn't need to care.

GtkScrolledWindow in GTK3 forces a min-height on its vscrollbar,
which forces the whole GtkStack to be at least that high as well.
This is worked around by a custom stylesheet that applies a lower
min-height on MudInputView's vertical scrollbar.

The resizing behaviour with GtkTextView+GtkScrolledView is buggy,
so the resizing is a bit erratic. This is to be investigated deeper
and improved still. It seems GtkTextView doesn't request a higher
size after a new line is started until some letters are written in
it, which is causing trouble and jumpy behaviour.